refactor: remove automatic `CI` env detection
BREAKING CHANGE: Previously when running with `CI` env `--skip-changelog` and `--skip-tag` were
automatically enabled. This is no longer the case, because the new GitHub releasing feature
makes it more appealing to run this tool in CI. To restore the old functionality for your
pipelines please explicitly add these tags to your command line.